+++ Algorithms-MaxFLow/FlowNetworkGraph.h
@@ -6,12 +6,13 @@
 class FlowNetworkGraph : public WeightedDirectedGraph
 {
      private:
-	     float** m_CurrentFlowMatrix;
+	     int** m_CurrentFlowMatrix;
 	     int m_S, m_T;
-	     float m_MaxFlow;
+	     int m_MaxFlow;
 
-          float GetResidualCapacityOfPath(const WeightedDirectedGraph& residualGraph, int* parentArrPath);
-          void UpdateResidualGraphAndFlow(int* parentArrPath, float residualCapacityOfPath, WeightedDirectedGraph& residualGraph);
+          //Service functions 
+          int GetResidualCapacityOfPath(const WeightedDirectedGraph& residualGraph, int* parentArrPath);
+          void UpdateResidualGraphAndFlow(int* parentArrPath, int residualCapacityOfPath, WeightedDirectedGraph& residualGraph);
           void PrintFulkersonUsingBfsOutput(bool* visitedArr, int numOfItr);
 
 
@@ -20,9 +21,9 @@ class FlowNetworkGraph : public WeightedDirectedGraph
           FlowNetworkGraph(int n,int s,int t);
 	     virtual ~FlowNetworkGraph();
 	     void MakeEmptyFlow();
-	     void AddFlow(int u, int v, float flowToAdd);
-	     bool IsFlowToAddIsValid(int u, int v, float flowToAdd);
-	     float GetResidualFlow(int u, int v);
+	     void AddFlow(int u, int v, int flowToAdd);
+	     bool IsFlowToAddIsValid(int u, int v, int flowToAdd);
+	     int GetResidualFlow(int u, int v);
           void FordFulkersonUsingBfs(); // Updating MaxFlow
           void PrintVertexesByBoolean(bool* visitedArr, string vertexNameStr, bool areVertexesAccesibleFromS);
           void PrintMinCut(bool* visitedArr);
